+++ lib/ui/home/similar_verses/similar_verses_page.dart
@@ -8,13 +8,13 @@ import 'similar_verse_manager.dart';
 class SimilarVersesPage extends StatefulWidget {
   const SimilarVersesPage({
     super.key,
+    required this.root,
     required this.strongsCode,
-    required this.fontSize,
     required this.isRtl,
   });
 
+  final String? root;
   final String strongsCode;
-  final double fontSize;
   final bool isRtl;
 
   @override
@@ -24,52 +24,64 @@ class SimilarVersesPage extends StatefulWidget {
 class _SimilarVersesPageState extends State<SimilarVersesPage> {
   final manager = SimilarVerseManager();
 
+  TextDirection get _textDirection =>
+      widget.isRtl ? TextDirection.rtl : TextDirection.ltr;
+
   @override
   void initState() {
     super.initState();
     manager.init(widget.strongsCode);
   }
 
+  String get title {
+    if (widget.root == null) {
+      return widget.strongsCode;
+    }
+    return '${widget.root} (${widget.strongsCode})';
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text(widget.strongsCode)),
-      body: ValueListenableBuilder<List<Reference>>(
-        valueListenable: manager.similarVersesNotifier,
-        builder: (context, verseList, child) {
-          return ListView.builder(
-            itemCount: verseList.length + 1,
-            itemBuilder: (context, index) {
-              if (index == 0) {
-                return Center(
-                  child: Text(
-                    '${verseList.length}',
-                    style: TextStyle(
-                      color: Theme.of(
-                        context,
-                      ).textTheme.bodyLarge?.color?.withAlpha(100),
+      appBar: AppBar(title: Text(title)),
+      body: Directionality(
+        textDirection: _textDirection,
+        child: ValueListenableBuilder<List<Reference>>(
+          valueListenable: manager.similarVersesNotifier,
+          builder: (context, verseList, child) {
+            return ListView.builder(
+              itemCount: verseList.length + 1,
+              itemBuilder: (context, index) {
+                if (index == 0) {
+                  return Center(
+                    child: Text(
+                      '${verseList.length}',
+                      style: TextStyle(
+                        color: Theme.of(
+                          context,
+                        ).textTheme.bodyLarge?.color?.withAlpha(100),
+                      ),
                     ),
+                  );
+                }
+                final referenceIndex = index - 1;
+                final reference = verseList[referenceIndex];
+                final formattedReference = _formatReference(reference);
+                return VerseListItem(
+                  key: ValueKey(reference),
+                  verseContentFuture: manager.getVerseContent(
+                    reference,
+                    widget.strongsCode,
+                    Theme.of(context).colorScheme.primary,
+                    20.0,
                   ),
+                  formattedReference: formattedReference,
+                  textDirection: _textDirection,
                 );
-              }
-              final referenceIndex = index - 1;
-              final reference = verseList[referenceIndex];
-              final formattedReference = _formatReference(reference);
-              return VerseListItem(
-                key: ValueKey(reference),
-                verseContentFuture: manager.getVerseContent(
-                  reference,
-                  widget.strongsCode,
-                  Theme.of(context).colorScheme.primary,
-                  20.0,
-                ),
-                formattedReference: formattedReference,
-                textDirection:
-                    widget.isRtl ? TextDirection.rtl : TextDirection.ltr,
-              );
-            },
-          );
-        },
+              },
+            );
+          },
+        ),
       ),
     );
   }